requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
celery-singleton==0.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.2.7
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.4
django-formtools==2.4.1
django-otp==1.3.0
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-rename-app==0.1.7
django-taggit==5.0.1
django-timezone-field==6.1.0
django-two-factor-auth==1.15.5
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Faker==23.2.1
h11==0.14.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
kombu==5.3.4
MarkupSafe==2.1.5
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.26
Pillow==10.1.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
tornado==6.3.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uWSGI==2.0.24
vine==5.1.0
wcwidth==0.2.10
zope.interface==6.1
